fix: 合并修复
[FIX] Issue when opening a Method
When opening a Method, the previous version loaded a vi placed inside the Do.vi, a part from the Method Vi, which is the one that has to be loaded. This returned as a result a Message Warning and the Plug-In did not execute properly.

Issue fix by adding a for loop in "Execute Open Message Handle.vi".

https://github.com/AdrianaOliva-DMIntegrations/Open-Actor-Method-From-Message.git